Updated all the XML source files for DB create step: pulse kernels, pulse_designs, pulse_sequences, experiments, NOT metabolites.  Had to remove all refs to 'scion.duhs.duke.edu' and tried to replace with github.com refs.
Bugfix in some of the dialogs. Reworked a wxglade FlexGrid to have 0 x 2 row / col size and just named it the name used in the dialog rather than finding some named element and getting the containing sizer name. This saves a step in dialog setup, AND it fixed an issue where inserted items started with space 2 (since the space holder item had been in space 1).