VESPA - updated lots of wxglade files to Py 3.9 wxglade 1.0.5 versions. Removed some test code and print statements. Restored floatspin(multiplier) files to version 1.0.0 and later in wxGlade.
PULSE - bug hack, when PanelView1D set into shared Sizer with PanelView2D the 2D stays at default 20,20 size, but 1D gets set to 0,1 for some reason. This fails when axhline is added due to a Singular Value error for transform conversion. Can't seem to get wxGlade to finesse this, so I'm setting up code for now to force initial size to 20,20.  It works but not elegant.